backends/ngfx: integrate ngfx backend

+++ configure.py
@@ -36,7 +36,7 @@ import tarfile
 import urllib.request
 import venv
 import zipfile
-from multiprocessing import Pool
+from multiprocessing import Pool, cpu_count
 from subprocess import run
 
 _ROOTDIR = op.abspath(op.dirname(__file__))
@@ -64,17 +64,32 @@ _EXTERNAL_DEPS = dict(
         url="https://renderdoc.org/stable/@VERSION@/renderdoc_@VERSION@.tar.gz",
         sha256="c8ec16f7463266641e21b64f8e436a452a15105e4bd517bf114a9349d74cc02e",
     ),
+    shaderc=dict(
+        version='2020.5',
+        dst_file='shaderc-@VERSION@.tar.gz',
+        url='https://github.com/google/shaderc/archive/v@VERSION@.tar.gz',
+        sha256='e96d8cb208b796cecb9e6cce437c7d1116343158ef3ea26277eb13b62cf56834',
+    ),
+    ngfx=dict(
+
+        url='https://github.com/gopro/ngfx.git',
+        branch='ngfx-integ-0.0.1',
+    ),
 )
 
 
 def _get_external_deps(args):
     deps = ["sxplayer"]
     if _SYSTEM == "Windows":
-        deps.append("pkgconf")
+        deps += ['pkgconf', 'shaderc']
     if "gpu_capture" in args.debug_opts:
         if _SYSTEM not in {"Windows", "Linux"}:
             raise Exception(f"Renderdoc is not supported on {_SYSTEM}")
         deps.append(_RENDERDOC_ID)
+    if _SYSTEM == 'Darwin':
+        deps += ['shaderc']
+    if args.enable_ngfx_backend:
+        deps += ['ngfx']
     return {dep: _EXTERNAL_DEPS[dep] for dep in deps}
 
 
@@ -134,22 +149,34 @@ def _download_extract(dep_item):
 
     name, dep = dep_item
 
-    version = dep["version"]
+    version = dep['version'] if 'version' in dep else ''
     url = dep["url"].replace("@VERSION@", version)
-    chksum = dep["sha256"]
+    chksum = dep.get("sha256")
+    is_git_repo = url.endswith('.git')
+    if is_git_repo:
+        git_branch = dep['branch'] if 'branch' in dep else 'main'
     dst_file = dep.get("dst_file", op.basename(url)).replace("@VERSION@", version)
+    if dst_file.endswith('.git'):
+        dst_file = dst_file[:-4]
     dst_base = op.join(_ROOTDIR, "external")
     dst_path = op.join(dst_base, dst_file)
     os.makedirs(dst_base, exist_ok=True)
 
     # Download
-    if not op.exists(dst_path) or not _file_chk(dst_path, chksum):
-        logging.info("downloading %s to %s", url, dst_file)
-        urllib.request.urlretrieve(url, dst_path)
-        assert _file_chk(dst_path, chksum)
+    if is_git_repo:
+        if not op.exists(dst_path):
+            logging.info('cloning %s', url)
+            run(['git', 'clone', url, '-b', git_branch, dst_path], check=True)
+    else:
+        if not op.exists(dst_path) or not _file_chk(dst_path, chksum):
+            logging.info("downloading %s to %s", url, dst_file)
+            urllib.request.urlretrieve(url, dst_path)
+            assert _file_chk(dst_path, chksum)
 
     # Extract
-    if tarfile.is_tarfile(dst_path):
+    if is_git_repo:
+        pass
+    elif tarfile.is_tarfile(dst_path):
         with tarfile.open(dst_path) as tar:
             dirs = {f.name for f in tar.getmembers() if f.isdir()}
             extract_dir = op.join(dst_base, _guess_base_dir(dirs))
@@ -169,16 +196,21 @@ def _download_extract(dep_item):
 
     # Remove previous link if needed
     target = op.join(dst_base, name)
-    rel_extract_dir = op.basename(extract_dir)
-    if op.islink(target) and os.readlink(target) != rel_extract_dir:
-        logging.info("unlink %s target", target)
-        os.unlink(target)
-    elif op.exists(target) and not op.islink(target):
-        logging.info("remove previous %s copy", target)
-        _rmtree(target)
+    if is_git_repo:
+        pass
+    else:
+        rel_extract_dir = op.basename(extract_dir)
+        if op.islink(target) and os.readlink(target) != rel_extract_dir:
+            logging.info("unlink %s target", target)
+            os.unlink(target)
+        elif op.exists(target) and not op.islink(target):
+            logging.info("remove previous %s copy", target)
+            _rmtree(target)
 
     # Link (or copy)
-    if not op.exists(target):
+    if is_git_repo:
+        pass
+    elif not op.exists(target):
         logging.info("symlink %s -> %s", target, rel_extract_dir)
         try:
             os.symlink(rel_extract_dir, target)
@@ -205,7 +237,28 @@ def _block(name, prerequisites=None):
 
     return real_decorator
 
+def _get_cmake_setup_options(cfg, build_type = '$(CMAKE_BUILD_TYPE)', generator = '$(CMAKE_GENERATOR)', prefix = '$(PREFIX)', external_dir = '$(EXTERNAL_DIR)'):
+    opts = f'-DCMAKE_BUILD_TYPE={build_type} -G\"{generator}\" -DCMAKE_INSTALL_PREFIX=\"{prefix}\" -DEXTERNAL_DIR=\"{external_dir}\"'
+    if _SYSTEM == 'Windows':
+        opts += f' -DCMAKE_INSTALL_INCLUDEDIR=Include -DCMAKE_INSTALL_LIBDIR=Lib -DCMAKE_INSTALL_BINDIR=Scripts'
+        # Always use MultiThreadedDLL (/MD), not MultiThreadedDebugDLL (/MDd)
+        # Some external libraries are only available in Release mode, not in Debug mode
+        # MSVC toolchain doesn't allow mixing libraries built in /MD mode with libraries built in /MDd mode
+        opts += f' -DCMAKE_MSVC_RUNTIME_LIBRARY:STRING=MultiThreadedDLL'
+        # Set Windows SDK Version
+        opts += f' -DCMAKE_SYSTEM_VERSION=$(CMAKE_SYSTEM_VERSION)'
+        # Set VCPKG Directory
+        opts += f' -DVCPKG_DIR=$(VCPKG_DIR)'
+    return opts
 
+def _get_cmake_compile_options(cfg, build_type = '$(CMAKE_BUILD_TYPE)', num_threads = cpu_count() + 1):
+    opts = f'--config {build_type} -j{num_threads}'
+    if cfg.args.verbose:
+        opts += '-v'
+    return opts
+
+def _get_cmake_install_options(cfg, build_type = '$(CMAKE_BUILD_TYPE)'):
+    return f'--config {build_type}'
 def _meson_compile_install_cmd(component, external=False):
     builddir = op.join("external", component, "builddir") if external else op.join("builddir", component)
     return ["$(MESON) " + _cmd_join(action, "-C", builddir) for action in ("compile", "install")]
@@ -241,17 +294,78 @@ def _renderdoc_install(cfg):
     renderdoc_dll = op.join(cfg.externals[_RENDERDOC_ID], "renderdoc.dll")
     return [f"copy {renderdoc_dll} {cfg.bin_path}"]
 
+@_block('shaderc-install')
+def _shaderc_install(cfg):
+    shaderc_cmake_setup_options = _get_cmake_setup_options(cfg, build_type='Release', generator='Ninja')
+    shaderc_cmake_compile_options = _get_cmake_compile_options(cfg, build_type='Release')
+    shaderc_cmake_install_options = _get_cmake_install_options(cfg, build_type='Release')
+    if _SYSTEM == 'Darwin':
+        shaderc_lib_filename = 'libshaderc_shared.1.dylib'
+    cmd = []
+    if _SYSTEM in ['Darwin', 'Windows']:
+        cmd += [ 'cd external/shaderc && python ./utils/git-sync-deps' ]
+        cmd += [
+            f'$(CMAKE) -S external/shaderc -B $(BUILDDIR)/shaderc {shaderc_cmake_setup_options} -DSHADERC_SKIP_TESTS=ON -DLLVM_USE_CRT_DEBUG=MD -DLLVM_USE_CRT_RELEASE=MD',
+            f'$(CMAKE) --build $(BUILDDIR)/shaderc {shaderc_cmake_compile_options}',
+            f'$(CMAKE) --install $(BUILDDIR)/shaderc {shaderc_cmake_install_options}'
+        ]
+        if _SYSTEM == 'Darwin':
+            cmd += [ f'install_name_tool -id @rpath/{shaderc_lib_filename} $(PREFIX)/lib/{shaderc_lib_filename}' ]
+    return cmd
 
+@_block('ngl-debug-tools-install')
+def _ngl_debug_tools_install(cfg):
+    ngl_debug_tools_cmake_setup_options = _get_cmake_setup_options(cfg)
+    ngl_debug_tools_cmake_compile_options = _get_cmake_compile_options(cfg)
+    ngl_debug_tools_cmake_install_options = _get_cmake_install_options(cfg)
+    cmd = [
+        f'$(CMAKE) -S ngl-debug-tools -B $(BUILDDIR)/ngl-debug-tools {ngl_debug_tools_cmake_setup_options}',
+        f'$(CMAKE) --build $(BUILDDIR)/ngl-debug-tools --verbose {ngl_debug_tools_cmake_compile_options}',
+        f'$(CMAKE) --install $(BUILDDIR)/ngl-debug-tools {ngl_debug_tools_cmake_install_options}'
+    ]
+    return cmd
+
+@_block('ngfx-install', [_shaderc_install])
+def _ngfx_install(cfg):
+    ngfx_deps = ['nlohmann-json', 'stb']
+    if _SYSTEM == 'Windows':
+        ngfx_deps += ['glm', 'd3dx12', 'glfw']
+    ngfx_deps_str = ' '.join(s for s in ngfx_deps)
+    if _SYSTEM == 'Windows':
+        ngfx_install_deps_cmd = f'echo& set OS={_SYSTEM}& set PKGS={ngfx_deps_str}& python external/ngfx/build_scripts/install_deps.py&'
+    else:
+        ngfx_install_deps_cmd = f'export OS={_SYSTEM} && export PKGS="{ngfx_deps_str}" && python external/ngfx/build_scripts/install_deps.py'
+    cmd = [ ngfx_install_deps_cmd ]
+    ngfx_cmake_setup_options = _get_cmake_setup_options(cfg)
+    ngfx_cmake_compile_options = _get_cmake_compile_options(cfg)
+    ngfx_cmake_install_options = _get_cmake_install_options(cfg)
+    cmd += [
+        f'$(CMAKE) -S external/ngfx -B $(BUILDDIR)/ngfx {ngfx_cmake_setup_options} -D$(NGFX_GRAPHICS_BACKEND)=ON',
+        f'$(CMAKE) --build $(BUILDDIR)/ngfx --verbose {ngfx_cmake_compile_options}',
+        f'$(CMAKE) --install $(BUILDDIR)/ngfx {ngfx_cmake_install_options}'
+    ]
+    if cfg.ngfx_graphics_backend == 'NGFX_GRAPHICS_BACKEND_DIRECT3D12':
+        cmd += [ '$(PREFIX)\\Scripts\\ngfx_compile_shaders_dx12.exe d3dBlitOp' ]
+    return cmd
 @_block("nodegl-setup", [_sxplayer_install])
-def _nodegl_setup(cfg):
-    nodegl_opts = []
-    if cfg.args.debug_opts:
-        debug_opts = ",".join(cfg.args.debug_opts)
-        nodegl_opts += [f"-Ddebug_opts={debug_opts}"]
 
-    if "gpu_capture" in cfg.args.debug_opts:
-        renderdoc_dir = cfg.externals[_RENDERDOC_ID]
-        nodegl_opts += [f"-Drenderdoc_dir={renderdoc_dir}"]
+def _nodegl_setup(cfg):
+    nodegl_setup_opts = ['--default-library', 'shared']
+    nodegl_debug_opts = []
+    if cfg.args.debug_opts:
+        debug_opts = ','.join(cfg.args.debug_opts)
+        nodegl_debug_opts += [f'-Ddebug_opts={debug_opts}']
+        if 'gpu_capture' in cfg.args.debug_opts:
+            if _SYSTEM in ('Linux', 'Windows'):
+                renderdoc_dir = cfg.externals[_RENDERDOC_ID]
+                nodegl_debug_opts += [f'-Drenderdoc_dir={renderdoc_dir}']
+        nodegl_setup_opts += nodegl_debug_opts
+
+    if cfg.args.enable_ngfx_backend:
+        nodegl_setup_opts += [f'-Denable_ngfx_backend=true']
+        nodegl_setup_opts += [f'-Dngfx_graphics_backend=$(NGFX_GRAPHICS_BACKEND)']
+    else:
+        nodegl_setup_opts += [f'-Denable_ngfx_backend=false']
 
     extra_library_dirs = []
     extra_include_dirs = []
@@ -273,13 +387,13 @@ def _nodegl_setup(cfg):
 
     if extra_library_dirs:
         opts = ",".join(extra_library_dirs)
-        nodegl_opts += [f"-Dextra_library_dirs={opts}"]
+        nodegl_setup_opts += [f"-Dextra_library_dirs={opts}"]
 
     if extra_include_dirs:
         opts = ",".join(extra_include_dirs)
-        nodegl_opts += [f"-Dextra_include_dirs={opts}"]
+        nodegl_setup_opts += [f"-Dextra_include_dirs={opts}"]
 
-    return ["$(MESON_SETUP) -Drpath=true " + _cmd_join(*nodegl_opts, "libnodegl", op.join("builddir", "libnodegl"))]
+    return ["$(MESON_SETUP) -Drpath=true " + _cmd_join(*nodegl_setup_opts, "libnodegl", op.join("builddir", "libnodegl"))]
 
 
 @_block("nodegl-install", [_nodegl_setup])
@@ -402,9 +516,10 @@ def _clean_py(cfg):
 @_block("clean", [_clean_py])
 def _clean(cfg):
     return [
-        _rd(op.join("builddir", "libnodegl")),
-        _rd(op.join("builddir", "ngl-tools")),
-        _rd(op.join("builddir", "tests")),
+        _rd(op.join("$(BUILDDIR)", "libnodegl")),
+        _rd(op.join("$(BUILDDIR)", "ngl-tools")),
+        _rd(op.join("$(BUILDDIR)", "tests")),
+         _rd(op.join('$(BUILDDIR)', 'shaderc')),
         _rd(op.join("external", "pkgconf", "builddir")),
         _rd(op.join("external", "sxplayer", "builddir")),
     ]
@@ -462,7 +577,9 @@ def _get_make_vars(cfg):
     # mechanism.
     #
     meson = "MAKEFLAGS= meson" if _SYSTEM != "Windows" else "meson"
-
+    cmake = 'cmake'
+    builddir = os.getenv("BUILDDIR", 'builddir')
+    buildtype = 'debug' if debug else 'release'
     meson_setup = [
         "setup",
         "--prefix",
@@ -470,29 +587,53 @@ def _get_make_vars(cfg):
         "--pkg-config-path",
         cfg.pkg_config_path,
         "--buildtype",
-        "debugoptimized" if debug else "release",
+        buildtype,
     ]
     if cfg.args.coverage:
         meson_setup += ["-Db_coverage=true"]
     if _SYSTEM != "MinGW" and "debug" not in cfg.args.buildtype:
         meson_setup += ["-Db_lto=true"]
 
+    cmake_builddtype = 'Debug' if debug else 'Release'
+    cmake_setup = [
+        f'-DCMAKE_INSTALL_PREFIX={cfg.prefix}',
+        f'-DCMAKE_BUILD_TYPE={cmake_builddtype}',
+        '-GNinja',
+    ]
     if _SYSTEM == "Windows":
         meson_setup += ["--bindir=Scripts", "--libdir=Lib", "--includedir=Include"]
     elif op.isfile("/etc/debian_version"):
         # Workaround Debian/Ubuntu bug; see https://github.com/mesonbuild/meson/issues/5925
         meson_setup += ["--libdir=lib"]
 
+    if _SYSTEM == 'Windows':
+        # Use Multithreaded DLL runtime library for debug and release configurations
+        # Third party libs are compiled in release mode
+        # Visual Studio toolchain requires all libraries to use the same runtime library
+        meson_setup += ['-Db_vscrt=md']
     ret = dict(
+        BUILDDIR = builddir,
+        PREFIX = cfg.prefix, #.replace('\\','/'),
         PIP=_cmd_join(python, "-m", "pip"),
         MESON=meson,
+        CMAKE = cmake,
+        CMAKE_GENERATOR = cfg.cmake_generator,
+        CMAKE_BUILD_TYPE = cfg.cmake_build_type,
     )
 
     ret["MESON_SETUP"] = "$(MESON) " + _cmd_join(*meson_setup, f"--backend={cfg.args.build_backend}")
     # Our tests/meson.build logic is not well supported with the VS backend so
     # we need to fallback on Ninja
-    ret["MESON_SETUP_TESTS"] = "$(MESON) " + _cmd_join(*meson_setup, "--backend=ninja")
+    meson_setup_tests = "$(MESON) " + _cmd_join(*meson_setup, "--backend=ninja")
+    ret["MESON_SETUP_TESTS"] = meson_setup_tests
 
+    ret['CMAKE_SETUP'] = 'cmake ' + _cmd_join(*cmake_setup)
+
+    if _SYSTEM == 'Windows':
+        ret['CMAKE_SYSTEM_VERSION'] = cfg.cmake_system_version
+        ret['VCPKG_DIR'] = cfg.args.vcpkg_dir
+    if cfg.args.enable_ngfx_backend:
+        ret['NGFX_GRAPHICS_BACKEND'] = cfg.ngfx_graphics_backend
     return ret
 
 
@@ -579,20 +720,35 @@ class _Config:
             self.pkg_config_path = op.join(self.prefix, "lib", "pkgconfig")
         self.externals = _fetch_externals(args)
 
+        if args.enable_ngfx_backend:
+            self.ngfx_graphics_backend = os.getenv('NGFX_GRAPHICS_BACKEND',
+                'NGFX_GRAPHICS_BACKEND_' + {'Windows': 'DIRECT3D12', 'Linux': 'VULKAN', 'Darwin': 'METAL'}[_SYSTEM])
+
+        self.cmake_generator = os.getenv("CMAKE_GENERATOR", {'Windows': 'Visual Studio 16 2019', 'Linux': 'Ninja', 'Darwin': 'Xcode', 'MinGW': 'Ninja'}[_SYSTEM])
         if _SYSTEM == "Windows":
-            _sxplayer_setup.prerequisites.append(_pkgconf_install)
-            if "gpu_capture" in args.debug_opts:
-                _nodegl_setup.prerequisites.append(_renderdoc_install)
+             # Set Windows SDK Version
+            self.cmake_system_version = os.getenv('CMAKE_SYSTEM_VERSION', '10.0.22000.0')
 
-            vcpkg_bin = op.join(args.vcpkg_dir, "installed", "x64-windows", "bin")
-            for f in glob.glob(op.join(vcpkg_bin, "*.dll")):
-                logging.info("copy %s to venv/Scripts", f)
-                shutil.copy2(f, op.join("venv", "Scripts"))
+        if args.buildtype == 'debug':
+            self.cmake_build_type = 'Debug'
+        else:
+            self.cmake_build_type = 'Release'
+        _sxplayer_setup.prerequisites.append(_pkgconf_install)
+        if "gpu_capture" in args.debug_opts:
+            _nodegl_setup.prerequisites.append(_renderdoc_install)
 
+        vcpkg_bin = op.join(args.vcpkg_dir, "installed", "x64-windows", "bin")
+        for f in glob.glob(op.join(vcpkg_bin, "*.dll")):
+            logging.info("copy %s to venv/Scripts", f)
+            shutil.copy2(f, op.join("venv", "Scripts"))
+
+        if args.enable_ngfx_backend:
+            _nodegl_setup.prerequisites.append(_ngfx_install)
     def get_env(self):
         sep = ":" if _SYSTEM == "MinGW" else os.pathsep
         env = {}
         env["PATH"] = sep.join((self.bin_path, "$(PATH)"))
+        env['EXTERNAL_DIR'] = op.join(_ROOTDIR, 'external')
         env["PKG_CONFIG_PATH"] = self.pkg_config_path
         if _SYSTEM == "Windows":
             env["PKG_CONFIG_ALLOW_SYSTEM_LIBS"] = "1"
@@ -620,8 +776,13 @@ def _run():
     parser.add_argument(
         "--build-backend", choices=("ninja", "vs"), default=default_build_backend, help="Build backend to use"
     )
+    parser.add_argument('-v','--verbose', action='store_true',
+        help='Enable verbose output')
     if _SYSTEM == "Windows":
         parser.add_argument("--vcpkg-dir", default=r"C:\vcpkg", help="Vcpkg directory")
+
+    parser.add_argument('--enable-ngfx-backend', action='store_true',
+                        help='Enable NGFX Backend')
 
     args = parser.parse_args()
 
@@ -639,7 +800,10 @@ def _run():
         _nodegl_updatedoc,
         _nodegl_updatespecs,
         _nodegl_updateglwrappers,
+        _ngl_debug_tools_install,
     ]
+    if _SYSTEM == 'Darwin':
+        blocks += [_shaderc_install]
     if args.coverage:
         blocks += [_coverage_html, _coverage_xml]
     makefile = _get_makefile(cfg, blocks)
